requirements.txt
pip
requests
requests_cache
Flask
cassandra-driver
flask_httpauth
passlib
itsdangerous

